Rename and move to jsr namespace
